+++ requirement/src/main/java/fk/retail/ip/requirement/model/RequirementDownloadLineItem.java
@@ -1,7 +1,6 @@
 package fk.retail.ip.requirement.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.inject.Inject;
 import fk.retail.ip.requirement.internal.entities.Requirement;
 import javax.xml.bind.annotation.XmlRootElement;
 import lombok.Data;
@@ -106,7 +105,7 @@ public class RequirementDownloadLineItem {
             this.intransitQty += (openReqQty != null ? openReqQty : 0);
             this.forecast = req.getRequirementSnapshot().getForecast();
         }
-        this.quantity = req.getQuantity()!= null ? req.getQuantity():0;
+        this.quantity = req.getQuantity()!= null ? req.getQuantity().intValue():0;
         this.supplier = req.getSupplier();
         this.mrp = req.getMrp();
         this.app = req.getApp() != null ? req.getApp() :0;
